refactor: move AnnotationBadge
