+++ src/app/components/annotation/AnnotationBar.jsx
