+++ src/app/components/history/diffItems/FlagDiff.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { AnnotationBadge } from "../../header/filter/FilterPopup";
+import { AnnotationBadge } from "../../annotation/AnnotationBadge";
 import {
   getAnnotationColor,
   getAnnotationTitle
