+++ src/scss/annotation.scss
@@ -71,3 +71,16 @@
     right: 0;
   }
 }
+
+.annotation-badge {
+  padding: 3px 9px;
+  border-radius: 200px; // just big enough!
+  border: 1.5px solid;
+  font-size: 0.9em;
+  font-weight: 500;
+  opacity: 0.8;
+
+  &:hover {
+    opacity: 1;
+  }
+}
