+++ src/scss/main.scss
@@ -61,4 +61,4 @@
 "breadcrumbs",
 "toggle",
 "resetTableView",
-"annotationBar";
+"annotation";
